qgscategorizedsymbolrenderer: factor out common model part
And add them in `QgsTemplatedCategorizedRendererModel`.
