+++ python/PyQt6/core/auto_generated/symbology/qgscategorizedsymbolrenderer.sip.in
@@ -172,6 +172,8 @@ categories.
 #include "qgscategorizedsymbolrenderer.h"
 %End
   public:
+    typedef QgsRendererCategory Category;
+
     QgsCategorizedSymbolRenderer( const QString &attrName = QString(), const QgsCategoryList &categories = QgsCategoryList() );
 %Docstring
 Constructor for QgsCategorizedSymbolRenderer.
